requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.4.1
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
outcome==1.2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
whitenoise==6.2.0
wsproto==1.1.0
